// website/src/components/types.ts
export interface ImageData {
  url: string
  width: number
  height: number
  description?: string
  caption?: string
  author?: string
  focusPoint?: {x: number; y: number}
  format: string

  ogURL: string
  smallTeaserURL: string
  mediumTeaserURL: string
  largeURL: string
}